feat(model): Taking it all inside
