+++ module/model-sql/__index__.ts
@@ -4,3 +4,4 @@ export * from './src/connection/base';
 export * from './src/connection/decorator';
 export * from './src/dialect/base';
 export * from './src/util';
+export * from './src/types';
