+++ module/model-sql/support/test/query.ts
@@ -5,7 +5,7 @@ import { Suite, Test } from '@travetto/test';
 import { castTo } from '@travetto/runtime';
 import { BaseModelSuite } from '@travetto/model/support/test/base';
 
-import { VisitStack } from '../../src/util';
+import { VisitStack } from '../../src/types';
 import { SQLModelService } from '../../src/service';
 
 @Schema()
